I think I know the issue - see that with Picsart we have a way to change the original prompts and replace the placeholder and with Imagen we have a different way. And maybe when we want to display the images we search by the format from the Picsart prompt...but this is not ok! Cause what we need is this: no matter what image model we use to generate images, we shall display the last image the user generated for that scene (no matter the model). But please do not mix the scenes....so for scene 1, we shall have the corresponding image displayed...for scene 2, the corresponding image. etc!! And in step 5 we disaply the same images for each scene as in step 4 and the same list is used for rendering the video. So there is a connection between all the images. Maybe you need to think of a clever way to identify images...and think also that for future we want to display to the user the full history of image generated for that scene.
help me fix one issue with how we display the images in Step 4, from create story page. When I work with Imagen AI model, and I generate one image for one Scene, somehow other images for other scenes are changed...so there is a bug there! But please analyze fist the solution with Picsart AI ...cause that is working just fine!! There is an issue with how we save the images with Picasrt in the Firebase...I am not sure...please investigate and fix

Have you fixed these?